+++ esh-term.iife.js
@@ -1827,6 +1827,10 @@ var eshTerm = (() => {
     "src/process-shim.js"() {
       init_global_inject();
       cwd = "/home/web";
+      globalThis.__eshGetCwd = () => cwd;
+      globalThis.__eshSetCwd = (v2) => {
+        cwd = v2;
+      };
       listeners = {};
       process = {
         platform: "linux",